+++ crates/matrix-sdk/CHANGELOG.md
@@ -39,6 +39,9 @@ All notable changes to this project will be documented in this file.
     `ClientCredentials`
   - `Oidc::restore_registered_client()` must NOT be called after
     `Oidc::register_client()` anymore.
+- [**breaking**]: The `authentication::qrcode` module now reexports types from
+  `oauth2` rather than `openidconnect`. Some type names might have changed.
+  ([#4604](https://github.com/matrix-org/matrix-rust-sdk/pull/4604))
 
 ## [0.10.0] - 2025-02-04